requirements.txt
# 核心框架
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# 命令行工具
click>=8.0.0

# 机器学习和深度学习
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.35.0
accelerate>=0.24.0

# 图像处理
Pillow>=10.0.0
opencv-python-headless>=4.8.0
numpy>=1.24.0

# HTTP 和 API
requests>=2.31.0
httpx>=0.25.0

# 日志和工具
loguru>=0.7.0
tqdm>=4.64.0

# LaMa 模型支持
iopaint>=1.3.0

# 科学计算
scikit-image>=0.21.0
easydict>=1.10

# 其他工具
pathlib2>=2.3.0 